cw2/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
WTForms==3.1.0
Werkzeug==2.3.7
Markdown==3.5.1
Pygments==2.16.1
python-dotenv==1.0.0
email-validator==2.1.0


